requirements.txt
python-telegram-bot[webhooks,job-queue]==20.7
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
redis[hiredis]==5.0.1
httpx==0.26.0
pydantic-settings==2.1.0
openai==1.12.0
python-dotenv==1.0.0
uvicorn==0.27.0
aiofiles==23.2.1
alembic==1.13.1
greenlet==3.0.3
psycopg2-binary==2.9.9
gunicorn